The first step towards separating SCSI request processing
Today, tcmu-runner owns the worker threads, thus forcing the users' code into the
plugin structure. This certainly works for minimal examples, but prevents any kind
of embedding into larger 3rd-party applications.

Production-grade applications generally have their own threading structure and want
to deal with socket/file/tcmu notifications in an async way. As such, I just need a
minimal library that handles device opening and command parsing.

There are several API sets now:
 - the API each plugin must implement (unchanged, see tcmu_handler)
 - the API available to the plugin (api.c)
 - the API the host process must implement (debug functions today)
 - the minimal API tcmu_lib will provide (a new concept, builds on api.c)
