+++ CMakeLists.txt
@@ -31,11 +31,15 @@ find_library(KMOD kmod)
 find_library(GFAPI gfapi)
 
 # Stuff for building the main binary
+add_library(tcmu
+  api.c
+  libtcmu.c
+  )
 add_executable(tcmu-runner
   main.c
-  api.c
   tcmuhandler-generated.c
   )
+target_link_libraries(tcmu-runner tcmu)
 target_compile_options(tcmu-runner
   PUBLIC -fPIC -Wl,-E
   )
